Added a new DocumentModel
