+++ package.json
@@ -45,19 +45,26 @@
     "watch:src": "tsc -w"
   },
   "dependencies": {
-    "@jupyterlab/application": "^3.0.6",
-    "@jupyterlab/apputils": "^3.0.5",
-    "@jupyterlab/coreutils": "^5.0.3",
-    "@jupyterlab/docregistry": "^3.0.6",
-    "@jupyterlab/filebrowser": "^3.0.6",
-    "@jupyterlab/launcher": "^3.0.5",
-    "@jupyterlab/mainmenu": "^3.0.5",
+    "@jupyterlab/application": "^3.1.0-alpha.7",
+    "@jupyterlab/apputils": "^3.1.0-alpha.7",
+    "@jupyterlab/coreutils": "^5.1.0-alpha.7",
+    "@jupyterlab/docregistry": "^3.1.0-alpha.7",
+    "@jupyterlab/filebrowser": "^3.1.0-alpha.7",
+    "@jupyterlab/launcher": "^3.1.0-alpha.7",
+    "@jupyterlab/mainmenu": "^3.1.0-alpha.7",
+    "@jupyterlab/observables": "^4.1.0-alpha.7",
+    "@jupyterlab/services": "~6.1.0-alpha.7",
+    "@jupyterlab/shared-models": "^3.1.0-alpha.7",
+    "@jupyterlab/ui-components": "^3.1.0-alpha.7",
+    "@lumino/commands": "^1.12.0",
     "@lumino/coreutils": "^1.5.3",
-    "@lumino/widgets": "^1.18.0"
+    "@lumino/signaling": "^1.4.3",
+    "@lumino/widgets": "^1.19.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1"
   },
   "devDependencies": {
     "@jupyterlab/builder": "^3.0.5",
-    "@lumino/messaging": "^1.4.3",
     "@typescript-eslint/eslint-plugin": "^2.27.0",
     "@typescript-eslint/parser": "^2.27.0",
     "eslint": "^7.5.0",
