+++ src/widget.ts
@@ -1,51 +1,99 @@
-import { Widget } from '@lumino/widgets';
+// Copyright 2018 Wolf Vollprecht
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-import { Signal, ISignal } from '@lumino/signaling';
+import { DocumentRegistry, DocumentWidget } from '@jupyterlab/docregistry';
 
-import { PromiseDelegate } from '@lumino/coreutils';
+import { MainMenu, JupyterLabMenu } from '@jupyterlab/mainmenu';
 
-/*
-	This is a typing-only import. If you use it directly, the mxgraph content
-  will be included in the main JupyterLab js bundle.
-*/
-// @ts-ignore
-import * as MX from './drawio/modulated.js';
+import { IChangedArgs, PathExt } from '@jupyterlab/coreutils';
 
-import './drawio/css/common.css';
+import { undoIcon, redoIcon } from '@jupyterlab/ui-components';
 
-import './drawio/styles/grapheditor.css';
+import { CommandRegistry } from '@lumino/commands';
 
-import { grapheditorTxt, defaultXml } from './pack';
+import { Signal } from '@lumino/signaling';
 
-const w = window as any;
-const webPath =
-  'https://raw.githubusercontent.com/jgraph/mxgraph/master/javascript/examples/grapheditor/www/';
+import { DrawIODocumentModel } from './model';
 
-w.RESOURCES_BASE = webPath + 'resources/';
-w.STENCIL_PATH = webPath + 'stencils/';
-w.IMAGE_PATH = webPath + 'images/';
-w.STYLE_PATH = webPath + 'styles/';
-w.CSS_PATH = webPath + 'styles/';
-w.OPEN_FORM = webPath + 'open.html';
+import { DrawIOWidget } from './panel';
 
-// w.mxBasePath = "http://localhost:8000/src/mxgraph/javascript/src/";
-//w.imageBasePath = 'http://localhost:8000/src/mxgraph/javascript/';
-w.mxLoadStylesheets = false; // disable loading stylesheets
-w.mxLoadResources = false;
+import { DrawIOToolbarButton } from './toolbar';
 
-/**
- * A DrawIO layout.
- */
-export class DrawIOWidget extends Widget {
-  /**
-   * Construct a `GridStackLayout`.
-   *
-   * @param info - The `DashboardView` metadata.
-   */
-  constructor() {
-    super();
-    void Private.ensureMx().then(mx => this._loadDrawIO(mx));
-    //this._loadDrawIO(MX);
+import {
+  formatPanelIcon,
+  plusIcon,
+  zoominIcon,
+  zoomoutIcon,
+  deleteIcon,
+  toFrontIcon,
+  toBackIcon,
+  fillColorIcon,
+  strokeColorIcon,
+  shadowIcon
+} from './icons';
+
+const DIRTY_CLASS = 'jp-mod-dirty';
+
+export class DrawIODocumentWidget extends DocumentWidget<DrawIOWidget, DrawIODocumentModel> {
+  constructor(options: DrawIODocumentWidget.IOptions<DrawIOWidget, DrawIODocumentModel>) {
+    super(options);
+    // Adding the buttons to the widget toolbar
+    // Modify containers style: line ~92700
+    // modulated.js, line: 84246
+    // actions: modulated.js line: ~93000
+    // actions: modulated.js line: ~111000
+
+    this._commands = options.commands;
+    this._menubar = new MainMenu(this._commands);
+
+    this._menubar.clearMenus();
+    this.toolbar.addClass('dio-toolbar');
+
+    //TODO:
+    // Add toolbar actions to change the default style of arrows and conections.
+    this._menuView = new JupyterLabMenu({ commands: this._commands });
+    this._menuView.menu.title.caption = 'View (Space+Drag to Scroll)';
+    this._menuView.menu.title.icon = formatPanelIcon;
+    this._menubar.addMenu(this._menuView.menu, { rank: 1 });
+
+    this._menuZoom = new JupyterLabMenu({ commands: this._commands });
+    //TODO: Change label to a view percentage
+    this._menuZoom.menu.title.label = 'Zoom';
+    this._menuZoom.menu.title.caption = 'Zoom (Alt+Mousewheel)';
+    this._menubar.addMenu(this._menuZoom.menu, { rank: 2 });
+
+    this._menuInsert = new JupyterLabMenu({ commands: this._commands });
+    this._menuInsert.menu.title.caption = 'Insert';
+    this._menuInsert.menu.title.icon = plusIcon;
+    this._menubar.addMenu(this._menuInsert.menu, { rank: 2 });
+
+    this.context.ready.then(async value => {
+      await this.content.ready.promise;
+
+      this._onTitleChanged();
+      this._addToolbarItems();
+      this.content.setContent(this.context.model.toString());
+      this._handleDirtyStateNew();
+
+      this.context.pathChanged.connect(this._onTitleChanged, this);
+      this.context.model.contentChanged.connect(this._onContentChanged, this);
+      this.context.model.stateChanged.connect(
+        this._onModelStateChangedNew,
+        this
+      );
+      this.content.graphChanged.connect(this._saveToContext, this);
+    });
   }
 
   /**
@@ -53,419 +101,187 @@ export class DrawIOWidget extends Widget {
    */
   dispose(): void {
     Signal.clearData(this);
-    this._editor.destroy();
     super.dispose();
   }
 
   /**
    * A promise that resolves when the csv viewer is ready.
    */
-  get ready(): PromiseDelegate<void> {
-    return this._ready;
+  get ready(): Promise<void> {
+    return this.content.ready.promise;
   }
 
-  get graphChanged(): ISignal<this, string> {
-    return this._graphChanged;
+  public getSVG(): string {
+    return ''; //this.mx.mxUtils.getXml(this._editor.editor.graph.getSvg());
   }
 
-  get mx(): any {
-    return this._mx;
+  public getAction(action: string): any {
+    return this.content.getAction(action);
   }
 
-  get editor(): any {
-    return this._editor;
+  public execute(action: string): void {
+    this.content.getAction(action).funct();
   }
 
-  get graph(): any {
-    return this._editor.editor.graph;
-  }
-
-  getAction(action: string): any {
-    return this._editor.actions.actions[action];
-  }
-
-  getActions(): any {
-    return this._editor.actions.actions;
-  }
-
-  setContent(newValue: string): void {
-    if (this._editor === undefined) {
-      return;
-    }
-
-    const oldValue = this._mx.mxUtils.getXml(this._editor.editor.getGraphXml());
-
-    if (oldValue !== newValue && !this._editor.editor.graph.isEditing()) {
-      if (newValue.length) {
-        const xml = this._mx.mxUtils.parseXml(newValue);
-        this._editor.editor.setGraphXml(xml.documentElement);
-      }
-    }
-  }
-
-  //Direction
   public toggleCellStyles(flip: string): void {
-    let styleFlip = this._editor.mx.mxConstants.STYLE_FLIPH;
-    switch (flip) {
-      case 'flipH':
-        styleFlip = this._editor.mx.mxConstants.STYLE_FLIPH;
-        break;
-      case 'flipV':
-        styleFlip = this._editor.mx.mxConstants.STYLE_FLIPV;
-        break;
-    }
-    this._editor.graph.toggleCellStyles(styleFlip, false);
+    this.content.toggleCellStyles(flip);
   }
 
-  //Align
   public alignCells(align: string): void {
-    if (this._editor.graph.isEnabled()) {
-      let pos = this._editor.mx.mxConstants.ALIGN_CENTER;
-      switch (align) {
-        case 'alignCellsLeft':
-          pos = this._editor.mx.mxConstants.ALIGN_LEFT;
-          break;
-        case 'alignCellsCenter':
-          pos = this._editor.mx.mxConstants.ALIGN_CENTER;
-          break;
-        case 'alignCellsRight':
-          pos = this._editor.mx.mxConstants.ALIGN_RIGHT;
-          break;
-        case 'alignCellsTop':
-          pos = this._editor.mx.mxConstants.ALIGN_TOP;
-          break;
-        case 'alignCellsMiddle':
-          pos = this._editor.mx.mxConstants.ALIGN_MIDDLE;
-          break;
-        case 'alignCellsBottom':
-          pos = this._editor.mx.mxConstants.ALIGN_BOTTOM;
-          break;
-      }
-      this._editor.graph.alignCells(pos);
-    }
+    this.content.alignCells(align);
   }
 
-  //Distribute
-  //drawio:command/horizontal
-  //drawio:command/vertical
   public distributeCells(horizontal: boolean): void {
-    this._editor.graph.distributeCells(horizontal);
+    this.content.distributeCells(horizontal);
   }
 
-  //Layout
-  //drawio:command/horizontalFlow
-  //drawio:command/verticalFlow
   public layoutFlow(direction: string): void {
-    let directionFlow = this._editor.mx.mxConstants.DIRECTION_WEST;
-    switch (direction) {
-      case 'horizontalFlow':
-        directionFlow = this._editor.mx.mxConstants.DIRECTION_WEST;
-        break;
-      case 'verticalFlow':
-        directionFlow = this._editor.mx.mxConstants.DIRECTION_NORTH;
-        break;
-    }
-
-    const mxHierarchicalLayout = this._editor.mx.mxHierarchicalLayout;
-    const layout = new mxHierarchicalLayout(this._editor.graph, directionFlow);
-
-    this._editor.editor.executeLayout(() => {
-      const selectionCells = this._editor.graph.getSelectionCells();
-      layout.execute(
-        this._editor.graph.getDefaultParent(),
-        selectionCells.length === 0 ? null : selectionCells
-      );
-    }, true);
+    this.content.layoutFlow(direction);
   }
 
   public horizontalTree(): void {
-    let tmp = this._editor.graph.getSelectionCell();
-    let roots = null;
-
-    if (
-      tmp === null ||
-      this._editor.graph.getModel().getChildCount(tmp) === 0
-    ) {
-      if (this._editor.graph.getModel().getEdgeCount(tmp) === 0) {
-        roots = this._editor.graph.findTreeRoots(
-          this._editor.graph.getDefaultParent()
-        );
-      }
-    } else {
-      roots = this._editor.graph.findTreeRoots(tmp);
-    }
-
-    if (roots !== null && roots.length > 0) {
-      tmp = roots[0];
-    }
-
-    if (tmp !== null) {
-      const mxCompactTreeLayout = this._editor.mx.mxCompactTreeLayout;
-      const layout = new mxCompactTreeLayout(this._editor.graph, true);
-      layout.edgeRouting = false;
-      layout.levelDistance = 30;
-
-      this._promptSpacing(
-        layout.levelDistance,
-        this._editor.mx.mxUtils.bind(this, (newValue: any) => {
-          layout.levelDistance = newValue;
-
-          this._editor.editor.executeLayout(() => {
-            layout.execute(this._editor.graph.getDefaultParent(), tmp);
-          }, true);
-        })
-      );
-    }
+    this.content.horizontalTree();
   }
 
   public verticalTree(): void {
-    let tmp = this._editor.graph.getSelectionCell();
-    let roots = null;
-
-    if (
-      tmp === null ||
-      this._editor.graph.getModel().getChildCount(tmp) === 0
-    ) {
-      if (this._editor.graph.getModel().getEdgeCount(tmp) === 0) {
-        roots = this._editor.graph.findTreeRoots(
-          this._editor.graph.getDefaultParent()
-        );
-      }
-    } else {
-      roots = this._editor.graph.findTreeRoots(tmp);
-    }
-
-    if (roots !== null && roots.length > 0) {
-      tmp = roots[0];
-    }
-
-    if (tmp !== null) {
-      const mxCompactTreeLayout = this._editor.mx.mxCompactTreeLayout;
-      const layout = new mxCompactTreeLayout(this._editor.graph, false);
-      layout.edgeRouting = false;
-      layout.levelDistance = 30;
-
-      this._promptSpacing(
-        layout.levelDistance,
-        this._editor.mx.mxUtils.bind(this, (newValue: any) => {
-          layout.levelDistance = newValue;
-
-          this._editor.editor.executeLayout(() => {
-            layout.execute(this._editor.graph.getDefaultParent(), tmp);
-          }, true);
-        })
-      );
-    }
+    this.content.verticalTree();
   }
 
   public radialTree(): void {
-    let tmp = this._editor.graph.getSelectionCell();
-    let roots = null;
-
-    if (
-      tmp === null ||
-      this._editor.graph.getModel().getChildCount(tmp) === 0
-    ) {
-      if (this._editor.graph.getModel().getEdgeCount(tmp) === 0) {
-        roots = this._editor.graph.findTreeRoots(
-          this._editor.graph.getDefaultParent()
-        );
-      }
-    } else {
-      roots = this._editor.graph.findTreeRoots(tmp);
-    }
-
-    if (roots !== null && roots.length > 0) {
-      tmp = roots[0];
-    }
-
-    if (tmp !== null) {
-      const mxRadialTreeLayout = this._editor.mx.mxRadialTreeLayout;
-      const layout = new mxRadialTreeLayout(this._editor.graph, false);
-      layout.levelDistance = 80;
-      layout.autoRadius = true;
-
-      this._promptSpacing(
-        layout.levelDistance,
-        this._editor.mx.mxUtils.bind(this, (newValue: any) => {
-          layout.levelDistance = newValue;
-
-          this._editor.editor.executeLayout(() => {
-            layout.execute(this._editor.graph.getDefaultParent(), tmp);
-
-            if (!this._editor.graph.isSelectionEmpty()) {
-              tmp = this._editor.graph.getModel().getParent(tmp);
-
-              if (this._editor.graph.getModel().isVertex(tmp)) {
-                this._editor.graph.updateGroupBounds(
-                  [tmp],
-                  this._editor.graph.gridSize * 2,
-                  true
-                );
-              }
-            }
-          }, true);
-        })
-      );
-    }
+    this.content.radialTree();
   }
 
   public organic(): void {
-    const mxFastOrganicLayout = this._editor.mx.mxFastOrganicLayout;
-    const layout = new mxFastOrganicLayout(this._editor.graph);
-
-    this._promptSpacing(
-      layout.forceConstant,
-      this._editor.mx.mxUtils.bind(this, (newValue: any) => {
-        layout.forceConstant = newValue;
-
-        this._editor.editor.executeLayout(() => {
-          let tmp = this._editor.graph.getSelectionCell();
-
-          if (
-            tmp === null ||
-            this._editor.graph.getModel().getChildCount(tmp) === 0
-          ) {
-            tmp = this._editor.graph.getDefaultParent();
-          }
-
-          layout.execute(tmp);
-
-          if (this._editor.graph.getModel().isVertex(tmp)) {
-            this._editor.graph.updateGroupBounds(
-              [tmp],
-              this._editor.graph.gridSize * 2,
-              true
-            );
-          }
-        }, true);
-      })
-    );
+    this.content.organic();
   }
 
   public circle(): void {
-    const mxCircleLayout = this._editor.mx.mxCircleLayout;
-    const layout = new mxCircleLayout(this._editor.graph);
-
-    this._editor.editor.executeLayout(() => {
-      let tmp = this._editor.graph.getSelectionCell();
-
-      if (
-        tmp === null ||
-        this._editor.graph.getModel().getChildCount(tmp) === 0
-      ) {
-        tmp = this._editor.graph.getDefaultParent();
-      }
-
-      layout.execute(tmp);
-
-      if (this._editor.graph.getModel().isVertex(tmp)) {
-        this._editor.graph.updateGroupBounds(
-          [tmp],
-          this._editor.graph.gridSize * 2,
-          true
-        );
-      }
-    }, true);
+    this.content.circle();
   }
-
-  private _loadDrawIO(mx: Private.MX): void {
-    this._mx = mx;
-
-    // Adds required resources (disables loading of fallback properties, this can only
-    // be used if we know that all keys are defined in the language specific file)
-    this._mx.mxResources.loadDefaultBundle = false;
-
-    // Fixes possible asynchronous requests
-    this._mx.mxResources.parse(grapheditorTxt);
-    const oParser = new DOMParser();
-    const oDOM = oParser.parseFromString(defaultXml, 'text/xml');
-    const themes: any = new Object(null);
-    themes[(this._mx.Graph as any).prototype.defaultThemeName] =
-      oDOM.documentElement;
-
-    // Workaround for TS2351: Cannot use 'new' with an expression whose type lacks a call or construct signature
-    const Editor: any = this._mx.Editor;
-    this._editor = new this._mx.EditorUi(new Editor(false, themes), this.node);
-    this._editor.refresh();
-
-    this._editor.editor.graph.model.addListener(
-      this._mx.mxEvent.NOTIFY,
-      (sender: any, evt: any) => {
-        const changes: any[] = evt.properties.changes;
-        for (let i = 0; i < changes.length; i++) {
-          if (changes[i].root) {
-            return;
-          }
-        }
-
-        if (this._editor.editor.graph.isEditing()) {
-          this._editor.editor.graph.stopEditing();
-        }
-
-        const graph = this._editor.editor.getGraphXml();
-        const xml = this._mx.mxUtils.getXml(graph);
-        this._graphChanged.emit(xml);
-      }
-    );
-
-    this._promptSpacing = this._mx.mxUtils.bind(
-      this,
-      (defaultValue: any, fn: any) => {
-        const FilenameDialog = this._mx.FilenameDialog;
-        const dlg = new FilenameDialog(
-          this._editor.editor,
-          defaultValue,
-          this._mx.mxResources.get('apply'),
-          (newValue: any) => {
-            fn(parseFloat(newValue));
-          },
-          this._mx.mxResources.get('spacing')
-        );
-        this._editor.editor.showDialog(dlg.container, 300, 80, true, true);
-        dlg.init();
-      }
-    );
-
-    this._ready.resolve(void 0);
-  }
-
-  private _editor: any;
-  private _mx: Private.MX;
-  private _promptSpacing: any;
-  private _ready = new PromiseDelegate<void>();
-  private _graphChanged = new Signal<this, string>(this);
-}
-
-/**
- * A namespace for module-level concerns like loading mxgraph
- */
-
-namespace Private {
-  export type MX = typeof MX;
-
-  let _mx: typeof MX;
-  let _mxLoading: PromiseDelegate<MX>;
 
   /**
-   * Asynchronously load the mx bundle, or return it if already available
+   * Handle a change to the title.
    */
-  export async function ensureMx(): Promise<MX> {
-    if (_mx) {
-      return _mx;
+  private _onTitleChanged(): void {
+    this.title.label = PathExt.basename(this.context.localPath);
+  }
+
+  private _onContentChanged(): void {
+    this.content.setContent(this.context.model.toString());
+  }
+
+  private _saveToContext(emiter: DrawIOWidget, content: string): void {
+    this.context.model.fromString(content);
+  }
+
+  private _onModelStateChangedNew(
+    sender: DocumentRegistry.IModel,
+    args: IChangedArgs<any>
+  ): void {
+    if (args.name === 'dirty') {
+      this._handleDirtyStateNew();
     }
+  }
 
-    if (_mxLoading) {
-      return await _mxLoading.promise;
+  private _handleDirtyStateNew(): void {
+    if (this.context.model.dirty) {
+      this.title.className += ` ${DIRTY_CLASS}`;
+    } else {
+      this.title.className = this.title.className.replace(DIRTY_CLASS, '');
     }
+  }
 
-    _mxLoading = new PromiseDelegate();
-    /*eslint-disable */
-    // @ts-ignore
-    _mx = await import('./drawio/modulated.js');
-    /*eslint-enable */
+  private _addToolbarItems(): void {
+    const actions = this.content.getActions();
 
-    _mxLoading.resolve(_mx);
-    return _mx;
+    // Menu view
+    this._menuView.addGroup([
+      { command: 'drawio:command/formatPanel' },
+      { command: 'drawio:command/outline' },
+      { command: 'drawio:command/layers' }
+    ]);
+
+    // Menu Zoom
+    this._menuZoom.addGroup([
+      { command: 'drawio:command/resetView' },
+      { command: 'drawio:command/fitWindow' },
+      { command: 'drawio:command/fitPageWidth' },
+      { command: 'drawio:command/fitPage' },
+      { command: 'drawio:command/fitTwoPages' },
+      { command: 'drawio:command/customZoom' }
+    ]);
+
+    // Menu insert
+    this._menuInsert.addGroup([
+      { command: 'drawio:command/insertLink' },
+      { command: 'drawio:command/insertImage' }
+    ]);
+
+    this.toolbar.addItem('ViewDropdown', this._menubar);
+
+    actions['zoomIn'].icon = zoominIcon;
+    actions['zoomIn'].tooltip = 'Zoom In (Ctrl+(Numpad)/Alt+Mousewheel)';
+    this.toolbar.addItem('zoomIn', new DrawIOToolbarButton(actions['zoomIn']));
+
+    actions['zoomOut'].icon = zoomoutIcon;
+    actions['zoomOut'].tooltip = 'Zoom Out (Ctrl-(Numpad)/Alt+Mousewheel)';
+    this.toolbar.addItem(
+      'zoomOut',
+      new DrawIOToolbarButton(actions['zoomOut'])
+    );
+
+    actions['undo'].icon = undoIcon;
+    actions['fillColor'].tooltip = 'Undo (Ctrl+Z)';
+    this.toolbar.addItem('undo', new DrawIOToolbarButton(actions['undo']));
+
+    actions['redo'].icon = redoIcon;
+    actions['redo'].tooltip = 'Redo (Ctrl+Shift+Z)';
+    this.toolbar.addItem('redo', new DrawIOToolbarButton(actions['redo']));
+
+    actions['delete'].icon = deleteIcon;
+    actions['delete'].tooltip = 'Delete';
+    this.toolbar.addItem('delete', new DrawIOToolbarButton(actions['delete']));
+
+    actions['toFront'].icon = toFrontIcon;
+    actions['toFront'].tooltip = 'To Front (Ctrl+Shift+F)';
+    this.toolbar.addItem(
+      'toFront',
+      new DrawIOToolbarButton(actions['toFront'])
+    );
+
+    actions['toBack'].icon = toBackIcon;
+    actions['toBack'].tooltip = 'To Back (Ctrl+Shift+B)';
+    this.toolbar.addItem('toBack', new DrawIOToolbarButton(actions['toBack']));
+
+    actions['fillColor'].icon = fillColorIcon;
+    actions['fillColor'].tooltip = 'Fill Color';
+    this.toolbar.addItem(
+      'fillColor',
+      new DrawIOToolbarButton(actions['fillColor'])
+    );
+
+    actions['strokeColor'].icon = strokeColorIcon;
+    actions['strokeColor'].tooltip = 'Fill Stroke Color';
+    this.toolbar.addItem(
+      'strokeColor',
+      new DrawIOToolbarButton(actions['strokeColor'])
+    );
+
+    actions['shadow'].icon = shadowIcon;
+    actions['shadow'].tooltip = 'Shadow';
+    this.toolbar.addItem('shadow', new DrawIOToolbarButton(actions['shadow']));
+  }
+
+  private _commands: CommandRegistry;
+  private _menubar: MainMenu;
+  private _menuView: JupyterLabMenu;
+  private _menuZoom: JupyterLabMenu;
+  private _menuInsert: JupyterLabMenu;
+}
+
+export namespace DrawIODocumentWidget {
+  export interface IOptions<T, U>
+    extends DocumentWidget.IOptions<DrawIOWidget, DrawIODocumentModel> {
+    commands: CommandRegistry;
   }
 }
